feat(installer): add certificat parameter on mac

+++ setup/macos/scripts/install.sh
@@ -3,11 +3,12 @@ WORKING_DIRECTORY=$(dirname "$(realpath "$0")")
 
 # Function to display usage information
 usage() {
-	echo "Usage: $0 [-h URL] [-u USERNAME] [-p PASSWORD] [-v LOG_LEVEL ] [-l] [-s] [-n] [-h]"
+	echo "Usage: $0 [-h URL] [-u USERNAME] [-p PASSWORD] [-v LOG_LEVEL ] [-c CERTIFICAT] [-s] [-n] [-h]"
 	echo "  -h HOST       host URL of the OCS Inventory NG server"
 	echo "  -u USERNAME   Username"
 	echo "  -p PASSWORD   Password"
 	echo "  -v LOG_LEVEL  Log level"
+	echo "  -c CERTIFICAT Path to the certificate file"
 	echo "  -s            Service mode (register service)"
 	echo "  -n            Run the agent now"
 	exit 1
@@ -18,12 +19,13 @@ SERVICE=false
 NOW=false # if true, we run the agent now with mode 2
 
 # Parse command-line arguments
-while getopts "h:u:p:v:lsnih" opt; do
+while getopts "h:u:p:v:c:snih" opt; do
 	case ${opt} in
 	h) URL=$OPTARG ;;
 	u) USERNAME=$OPTARG ;;
 	p) PASSWORD=$OPTARG ;;
 	v) LOG_LEVEL=$OPTARG ;;
+	c) CERTIFICAT=$OPTARG ;;
 	s) SERVICE=true ;;
 	n) NOW=true ;;
 	*) usage ;;
@@ -43,6 +45,17 @@ else
 	NOW="no"
 fi
 
+# Check if the required parameters are set
+if [ -z "$URL" ] || [ -z "$USERNAME" ] || [ -z "$PASSWORD" ] || [ -z "$LOG_LEVEL" ]; then
+	usage
+fi
+
+# Check if the certificate file exists
+if [ ! -f "$CERTIFICAT" ]; then
+	echo "Certificate file not found: $CERTIFICAT"
+	exit 1
+fi
+
 # Save the configuration to /tmp/installer_config_file.txt
 cat <<EOF | sudo tee "/tmp/installer_config.txt" >/dev/null
 server=${URL}
@@ -51,6 +64,7 @@ password=${PASSWORD}
 logLevel=${LOG_LEVEL}
 serviceMode=${SERVICE}
 runNow=${NOW}
+certificat=${CERTIFICAT}
 EOF
 
 # build the package path by navigating one level up from the script directory
